+++ packages/components/table/src/base/table-body.tsx
@@ -6,8 +6,14 @@ import {ReactNode} from "react";
 export interface TableBodyProps<T>
   extends TableBodyBaseProps<T>,
     Omit<HTMLNextUIProps<"tbody">, keyof TableBodyBaseProps<T>> {
-  /** Provides content to display when there are no rows in the table. */
-  renderEmptyState?: () => ReactNode;
+  /**
+   * Provides content to display a loading component when the `loadingState` is `loading` or `loadingMore`.
+   */
+  loadingContent?: ReactNode;
+  /**
+   *  Provides content to display when there are no rows in the table.
+   * */
+  emptyContent?: ReactNode;
 }
 
 const TableBody = TableBodyBase as <T>(props: TableBodyProps<T>) => JSX.Element;