+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @nextui-org/accordion
 
+## 0.0.0-dev-v2-20230423212921
+
+### Patch Changes
+
+- Table improvements
+- Updated dependencies
+  - @nextui-org/use-aria-accordion-item@0.0.0-dev-v2-20230423212921
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230423212921
+  - @nextui-org/shared-icons@0.0.0-dev-v2-20230423212921
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230423212921
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230423212921
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230423212921
+  - @nextui-org/system@0.0.0-dev-v2-20230423212921
+  - @nextui-org/theme@0.0.0-dev-v2-20230423212921
+
+## 0.0.0-dev-v2-20230423211529
+
+### Patch Changes
+
+- Table component addded
+- Updated dependencies
+  - @nextui-org/use-aria-accordion-item@0.0.0-dev-v2-20230423211529
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230423211529
+  - @nextui-org/shared-icons@0.0.0-dev-v2-20230423211529
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230423211529
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230423211529
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230423211529
+  - @nextui-org/system@0.0.0-dev-v2-20230423211529
+  - @nextui-org/theme@0.0.0-dev-v2-20230423211529
+
 ## 0.0.0-dev-v2-20230421004240
 
 ### Patch Changes