+++ packages/hooks/use-aria-toggle-button/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @nextui-org/use-aria-toggle-button
 
+## 0.0.0-dev-v2-20230423212921
+
+### Patch Changes
+
+- Table improvements
+- Updated dependencies
+  - @nextui-org/use-aria-button@0.0.0-dev-v2-20230423212921
+
+## 0.0.0-dev-v2-20230423211529
+
+### Patch Changes
+
+- Table component addded
+- Updated dependencies
+  - @nextui-org/use-aria-button@0.0.0-dev-v2-20230423211529
+
 ## 0.0.0-dev-v2-20230420134926
 
 ### Patch Changes